update upt category controller

+++ IMS/IMS.Models/Entity/ProductInvoiceInfo.cs
@@ -13,9 +13,9 @@ namespace IMS.Models.Entity
         public int StoreInfoId { get; set; }    
         public int CustomerInfoId { get; set; }
         public DateTime TransactionDate { get; set; }
-        public double NetAmount { get;set; }
-        public double DiscountAmount { get;set; }
-        public double TotalAmount { get; set; }
+        public float NetAmount { get;set; }
+        public float DiscountAmount { get;set; }
+        public float TotalAmount { get; set; }
         public string Remarks { get;set; }
         public int BillStatus { get; set; }
         public string CancellationRemarks { get; set; }
